Worked on calculations of totals
Created a reference class to references that are used in transferable
that might need to be used elsewhere. Added an inner class  for each
item to hold the object label, how many objects there are to transfer,
and if it can be transferred.  This is in preparation to replace the
current checkboxs with a table that shows the label, a checkbox, and
then the number of items of that type that will be transferred.